Move duplicated code into one class
