+++ src/lib/Query/Location/CriterionVisitor/IsContainer.php
@@ -6,29 +6,12 @@
  */
 namespace Ibexa\Solr\Query\Location\CriterionVisitor;
 
-use Ibexa\Contracts\Core\Repository\Values\Content\Query\Criterion;
-use Ibexa\Contracts\Core\Repository\Values\Content\Query\Criterion\Operator;
-use Ibexa\Contracts\Solr\Query\CriterionVisitor;
+use Ibexa\Solr\Query\Common\CriterionVisitor\CommonIsContainer;
 
-final class IsContainer extends CriterionVisitor
+final class IsContainer extends CommonIsContainer
 {
-    public function canVisit(Criterion $criterion): bool
+    public function getTargetField(): string
     {
-        return $criterion instanceof Criterion\IsContainer && $criterion->operator === Operator::EQ;
-    }
-
-    public function visit(Criterion $criterion, CriterionVisitor $subVisitor = null): string
-    {
-        $value = $criterion->value;
-
-        if (!is_array($value) || !is_bool($value[0])) {
-            throw new \LogicException(sprintf(
-                '%s value should be of type array<bool>, received %s.',
-                Criterion\IsContainer::class,
-                get_debug_type($value),
-            ));
-        }
-
-        return 'is_container_b:' . $this->toString($value[0]);
+        return 'is_container_b';
     }
 }
